requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apache-beam==2.43.0
apache-flink==1.17.0
apache-flink-libraries==1.17.0
argilla==1.3.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
avro-python3==1.9.2.1
backoff==2.2.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.25
clickhouse-connect==0.5.25
cloudpickle==2.2.0
coloredlogs==15.0.1
crcmod==1.7
dataclasses-json==0.5.7
Deprecated==1.2.13
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
docopt==0.6.2
docstring-parser==0.15
duckdb==0.8.0
et-xmlfile==1.1.0
fake-useragent==1.1.3
fastapi==0.92.0
fastavro==1.4.7
filelock==3.12.0
find-libpython==0.3.0
flatbuffers==23.5.9
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
gptcache==0.1.27
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
hdfs==2.7.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
joblib==1.2.0
kafka-python==2.0.2
langchain==0.0.176
langflow==0.0.78
llama-cpp-python==0.1.50
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.21.6
objsize==0.5.2
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.12
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pemja==0.3.0
pipreqs==0.4.13
posthog==3.0.1
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.6
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydot==1.4.2
pyflink==1.0
Pygments==2.15.1
pymongo==3.13.0
pypandoc==1.11
pypdf==3.9.0
pysrt==1.1.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
regex==2023.5.5
requests==2.30.0
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.25.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
tqdm==4.65.0
typer==0.7.0
types-PyYAML==6.0.12.10
typing-inspect==0.9.0
typing_extensions==4.5.0
unstructured==0.5.13
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.14.1
XlsxWriter==3.1.1
yarg==0.1.9
yarl==1.9.2
youtube-transcript-api==0.6.0
zstandard==0.21.0
